Implement Audio Manager
Added an audio manager class which will hopefully sort and manage all sounds in game, and give them a random volume and pitch
Also, I hope Rage Against The Machine is cool for a background sound clip, because it's either that or Creedence Clearwater Revival - Fortunate Son, I might implement both.
Think both set the mood of the game quite well.
Need to add some settings to disable and enable sound and stuff like that.
Reckon it shouldn't be too hard